Also move `formatZipCode`
It's pretty closely related to the rest of the address stuff, so even though it's not used in multiple packages, it arguably belongs in common/address.
