+++ loader/src/utils.js
@@ -2,7 +2,11 @@ const timers = require("node:timers/promises");
 const nodeUtil = require("util");
 const Sentry = require("@sentry/node");
 const got = require("got");
-const { matchableAddress, parseUsAddress } = require("univaf-common/address");
+const {
+  formatZipCode,
+  matchableAddress,
+  parseUsAddress,
+} = require("univaf-common/address");
 const config = require("./config");
 const {
   GraphQlError,
@@ -176,6 +180,8 @@ module.exports = {
 
   parseUsAddress,
 
+  formatZipCode,
+
   /**
    * Parse and return a US-style phone number with an area code and, optionally,
    * a country code. This handles some oddball situations like phone numbers
@@ -198,30 +204,6 @@ module.exports = {
     }
 
     throw new ParseError(`Invalid U.S. phone number: "${text}"`);
-  },
-
-  /**
-   * Format a US ZIP code into its canonical 5-digit form (plus an optional
-   * 4-digit suffix). This is useful since a lot of systems return ZIP codes
-   * without leading zeroes.
-   * @param {string|number} zipCode
-   * @returns {string}
-   */
-  formatZipCode(zipCode) {
-    if (typeof zipCode === "number" && !Number.isInteger(zipCode)) {
-      throw new SyntaxError(`ZIP code is not an integer: "${zipCode}"`);
-    }
-
-    const parts = zipCode.toString().split("-");
-    if (parts.length > 2) {
-      throw new SyntaxError(`ZIP code has multiple dashes: "${zipCode}"`);
-    }
-
-    let result = parts[0].padStart(5, "0");
-    if (parts.length > 1) {
-      result = `${result}-${parts[1].padStart(4, "0")}`;
-    }
-    return result;
   },
 
   /**
